requirements.txt
torch>=1.10.0
PyYAML>=6.0.0
scipy